Extract shared Moodle/HTML/time helpers to cut duplication
Move iso/cache/time helpers, Moodle AJAX unwrap, notice-board discovery,
header-region filtering, path removal, and argparse help formatting into
shared modules so providers and adapters stop copy-pasting identical helpers.
