+++ src/kaist_cli/core/fsutil.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+import shutil
+from pathlib import Path
+
+
+def remove_path(path: Path) -> None:
+    if path.is_symlink() or path.is_file():
+        path.unlink()
+        return
+    if path.is_dir():
+        shutil.rmtree(path)
