+++ src/kaist_cli/v2/parser.py
@@ -3,9 +3,7 @@ from __future__ import annotations
 import argparse
 from typing import Any, Callable
 
-
-class _HelpFormatter(argparse.ArgumentDefaultsHelpFormatter, argparse.RawDescriptionHelpFormatter):
-    pass
+from ..cli.help_format import HelpFormatter as _HelpFormatter
 
 
 Handler = Callable[[argparse.Namespace], Any]
